Add Vercel serverless setup
