+++ api/index.js
@@ -1,0 +1,4 @@
+const serverlessExpress = require('@vendia/serverless-express');
+const app = require('../backend/app');
+
+module.exports = serverlessExpress({ app });
