+++ backend/package.json
@@ -11,7 +11,8 @@
     "cors": "^2.8.5",
     "body-parser": "^1.20.2",
     "jsonwebtoken": "^9.0.0",
-    "multer": "^1.4.5-lts.1"
+    "multer": "^1.4.5-lts.1",
+    "@vendia/serverless-express": "^4.8.2"
   },
   "devDependencies": {
     "nodemon": "^2.0.22"
